requirements.txt
chromadb
discord.py
faster_whisper
huggingface_hub
PyAudio
python-dotenv
torch
webrtcvad-wheels
discord-ext-voice-recv
llama-cpp-python
requests
tenacity